backend/requirements.txt
# Backend-only dependencies (separate env from bot recommended).
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9
httpx>=0.27
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
pydantic-settings>=2.0
PyJWT>=2.8
celery[redis]>=5.4
redis[hiredis]>=5.0

